requirements.txt

telethon
pushbullet.py
